Implemented diagrams into docutils parser
This feature depends on asciitosvg[1] to operate. It work using the diagram
directive, and by drawing the diagram in ASCII art for example:

   .. diagram::

       #-----------#
       |   Foo !   +------'
       #-----------#      |
                          |        #-----------#
                          '------->|  BAR !!!  +---'
                                   #-----------#   |
                                          ^        |
                                          '--------'

Formatting rules can be found in asciitosvg project readme[2].

Two new options are now available in the docutils parser configuration:

- a2s_binary_path: path to the a2s binary file
- a2s_container_class: css class to use for the diagram container

If a2s_binary_path is omitted, diagrams will be silently hidden.

[1] https://9vx.org/~dho/a2s/
[2] https://bitbucket.org/dhobsd/asciitosvg