Change configuration filename
The configuration filename `config.yml` has been deprecated and the
preferred configuration filename is now `mynt.yml`.

If the configuration file is named `config.yml`, a warning is issued
informing the user it has been deprecated and should be renamed to
`mynt.yml`. Support for the configuration filename `config.yml` may be
removed in a future release.